requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.7.0
apache-log-parser==1.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
autopep8==2.3.1
beautifulsoup4==4.12.3
billiard==4.2.0
blinker==1.8.2
boto3==1.35.92
botocore==1.35.92
branca==0.8.0
Brotli==1.1.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
crispy-bootstrap4==2024.1
cron-descriptor==1.4.5
cryptography==43.0.3
cssselect2==0.7.0
daphne==4.1.2
distro==1.9.0
Django==4.2.19
django-appconf==1.0.6
django-axes==6.5.1
django-block-ip==0.2.0
django-celery-beat==2.7.0
django-ckeditor==6.7.2
django-cloudflare==0.0.1
django-compressor==4.5.1
django-cors-headers==4.6.0
django-crispy-forms==2.1
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-filer==3.3.0
django-formtools==2.5.1
django-js-asset==3.1.2
django-meta==2.4.2
django-polymorphic==3.1.0
django-pwa==2.0.1
django-seo-js==0.4.1
django-silk==5.2.0
django-timezone-field==7.0
easy-thumbnails==2.10
exceptiongroup==1.2.2
Flask==3.0.3
folium==0.18.0
fonttools==4.54.1
frozenlist==1.5.0
gevent==24.11.1
git-filter-repo==2.45.0
gprof2dot==2024.6.6
graphene==3.4.3
graphene-django==3.2.2
graphql-core==3.2.5
graphql-relay==3.2.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
http-ece==1.2.1
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipcalc==1.99.0
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.0
jmespath==1.0.1
kombu==5.4.1
Levenshtein==0.26.1
log-parser==2.3
lxml==5.2.2
lz4==4.3.3
MarkupSafe==3.0.1
msgpack==1.1.0
multidict==6.1.0
numpy==2.1.2
openai==1.57.0
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==10.3.0
promise==2.3
prompt_toolkit==3.0.47
propcache==0.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.10
py-vapid==1.9.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pydyf==0.11.0
pylibmc==1.6.3
pymemcache==4.0.0
pyOpenSSL==24.2.1
pyphen==0.16.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-Levenshtein==0.26.1
pytz==2025.1
pywebpush==2.0.3
PyYAML==6.0.2
RapidFuzz==3.11.0
rcssmin==1.1.2
redis==5.2.0
reportlab==4.2.0
requests==2.32.3
rjsmin==1.2.2
s3transfer==0.10.4
service-identity==24.2.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.0
svglib==1.5.1
text-unidecode==1.3
tinycss2==1.3.0
tomli==2.0.1
tqdm==4.67.1
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
ua-parser==0.18.0
urllib3==2.2.3
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==62.3
webencodings==0.5.1
Werkzeug==3.1.3
xyzservices==2024.9.0
yarl==1.18.0
zope.event==5.0
zope.interface==7.1.1
zopfli==0.2.3
